refactor(Payment): 결제 로직 추가를 위한 payment 도메인 추가
