+++ src/main/java/com/example/hot_deal/payment/domain/Payment.java
@@ -1,4 +1,4 @@
-package com.example.hot_deal.order.domain.payment;
+package com.example.hot_deal.payment.domain;
 
 import com.example.hot_deal.common.domain.BaseTimeEntity;
 import com.example.hot_deal.common.domain.Price;
